% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{anatomecha_splits}
\alias{anatomecha_splits}
\title{Splitsio data: anatomecha's splits}
\format{
An object of class \code{data.frame} with 22 rows and 2 columns.
}
\usage{
anatomecha_splits
}
\description{
See vignette splits
}
\seealso{
Other data: 
\code{\link{sio_df}},
\code{\link{src_run_df}}
}
\concept{data}
\keyword{datasets}
