% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{common_splits}
\alias{common_splits}
\title{Most recorded splits}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 19 rows and 2 columns.
}
\usage{
common_splits
}
\description{
See vignette splits; splits the most runners have recorded.
}
\keyword{datasets}
