% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cum_t_route_vis.R
\name{cum_t_route_vis}
\alias{cum_t_route_vis}
\title{Players take different routes}
\usage{
cum_t_route_vis(
  ylim = range(supermetroid::splits_df$realtime_end_ms)/1000/60,
  font_size = 20
)
}
\arguments{
\item{ylim}{Set y min and max (shiny slider input).}

\item{selected_splits}{String vector for \%in\% filter.}
}
\description{
Acts on \link{sio_df}.
}
