% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{distinct_routes}
\alias{distinct_routes}
\title{Distinct routes}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 211 rows and 3 columns.
}
\usage{
distinct_routes
}
\description{
Used in route-graph and route-matching vignettes
}
\seealso{
Other smdata: 
\code{\link{anatomecha_splits}},
\code{\link{sio_df}},
\code{\link{src_df}()}
}
\concept{smdata}
\keyword{datasets}
