% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranked_runs.R
\name{ranked_runs}
\alias{ranked_runs}
\title{Visualise each run}
\usage{
ranked_runs(selected_split = NULL, base_size = 40)
}
\arguments{
\item{base_size}{For sm_theme theme_minimal.
segment. If an \link{anatomecha_splits} split is specificied, highlight that.}

\item{selected_segment}{If NULL plot produces plot with no highlighted}
}
\description{
Takes sio_df. Need to join rankings to get y-axis ordering.
}
