% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route_graph.R
\name{route_graph}
\alias{route_graph}
\title{Graphing routes players take}
\usage{
route_graph(
  base_size = base_size,
  selected_splits = c("kraid", "ice beam", "gravity suit", "draygon", "ridley",
    "varia suit", "space jump")
)
}
\arguments{
\item{base_size}{for theme_minimal}

\item{selected_splits}{Vector of splits of interest.}

\item{segments}{Spltis of interest}
}
\description{
Graphing routes players take
}
