% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sio_df}
\alias{sio_df}
\title{Splitsio smdata: time series observations}
\format{
An object of class \code{data.frame} with 19815 rows and 11 columns.
}
\usage{
sio_df
}
\description{
See vignette splits; use \link{splits_df} for analysis.
}
\keyword{datasets}
