% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sio_routes_wide}
\alias{sio_routes_wide}
\title{Wide array of runtime at segment by run}
\format{
An object of class \code{grouped_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 538 rows and 24 columns.
}
\usage{
sio_routes_wide
}
\description{
splits.io routes
}
\keyword{datasets}
