% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sio_splits_df.R
\docType{data}
\name{sio_splits_df}
\alias{sio_splits_df}
\title{Splits.io data for segments}
\format{
An object of class \code{data.frame} with 19815 rows and 12 columns.
}
\usage{
sio_splits_df
}
\description{
Each row is a split time (segment) for a specific run.
}
\keyword{datasets}
