% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm_theme.R
\docType{data}
\name{sm_col_h}
\alias{sm_col_h}
\title{Super metroid colour heirarchy}
\format{
An object of class \code{list} of length 5.
}
\usage{
sm_col_h
}
\description{
Super metroid colour heirarchy
}
\keyword{datasets}
