% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm_theme.R
\docType{data}
\name{sm_cols}
\alias{sm_cols}
\title{Super Metroid visual theme elements}
\format{
An object of class \code{list} of length 13.
}
\usage{
sm_cols
}
\description{
Super Metroid visual theme elements
}
\keyword{datasets}
