% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{splits_df}
\alias{splits_df}
\title{Splits df with src info}
\format{
An object of class \code{data.frame} with 19815 rows and 16 columns.
}
\usage{
splits_df
}
\description{
Use this dataset for analysis
}
\keyword{datasets}
