% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{src_df}
\alias{src_df}
\title{Speedrun.com leaderboard smdata}
\format{
An object of class \code{data.frame} with 586 rows and 7 columns.
}
\usage{
src_df
}
\description{
Player and run data joined. See src vignette.
}
\keyword{datasets}
