% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{src_run_df}
\alias{src_run_df}
\title{Speedrun.com leaderboard data}
\format{
An object of class \code{data.frame} with 582 rows and 4 columns.
}
\usage{
src_run_df
}
\description{
Speedrun.com leaderboard data
}
\seealso{
Other data: 
\code{\link{anatomecha_splits}},
\code{\link{sio_df}}
}
\concept{data}
\keyword{datasets}
