% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{src_top_df}
\alias{src_top_df}
\title{Top games from speedrun.com}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 42 rows and 3 columns.
}
\usage{
src_top_df
}
\description{
Hand transcribed top five from each category across games
and for SNES.
}
\keyword{datasets}
