% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm_theme.R
\name{theme_sm}
\alias{theme_sm}
\title{Add Super Metroid theme to ggplot}
\usage{
theme_sm(base_size = 15)
}
\arguments{
\item{base_size}{for theme_minimal}

\item{leaderboard_api}{string for api "splits.io" or "speedrunner.com"

Stuff to do

Possibly add source as an argument}
}
\description{
Add Super Metroid theme to ggplot
}
