% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_run_raincloud.R
\name{all_run_raincloud}
\alias{all_run_raincloud}
\title{Raincloud plot of speedrun.com run times}
\usage{
all_run_raincloud(
  run_df = src_run_df,
  xmin = min(src_run_df$t_s)/60,
  xmax = max(src_run_df$t_s)/60,
  base_size = 20,
  source = "speedrun.com via srcomapi"
)
}
\arguments{
\item{run_df}{Dataframe of column of run times in seconds.}

\item{xmin}{Set limits of x axis, for shiny slider input, in minutes.}

\item{xmax}{Set max of x axis, shiny slider, in minutes.}

\item{base_size}{For theme minimal}

\item{source}{Caption provenance}
}
\description{
Raincloud plot of speedrun.com run times
}
