% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{from_to_df}
\alias{from_to_df}
\title{Dataframe for graphing routes}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2673 rows and 4 columns.
}
\usage{
from_to_df
}
\description{
splits.io routes in tidygraph form
}
\keyword{datasets}
