% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route_graph.R
\name{route_graph}
\alias{route_graph}
\title{Graphing routes players take}
\usage{
route_graph(
  font_size = 15,
  selected_splits = supermetroid::common_splits$game_event
)
}
\arguments{
\item{selected_splits}{Vector of splits of interest.}

\item{segments}{Spltis of interest}
}
\description{
Graphing routes players take
}
