% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sio_df}
\alias{sio_df}
\title{Splitsio smdata: time series observations}
\format{
An object of class \code{data.frame} with 19815 rows and 14 columns.
}
\usage{
sio_df
}
\description{
See vignette splits
}
\seealso{
Other smdata: 
\code{\link{anatomecha_splits}},
\code{\link{distinct_routes}},
\code{\link{from_to_df}},
\code{\link{sio_routes_wide}},
\code{\link{sio_splits_df}},
\code{\link{src_df}()}
}
\concept{smdata}
\keyword{datasets}
