% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sio_df}
\alias{sio_df}
\title{Datasets}
\format{
An object of class \code{data.frame} with 19815 rows and 14 columns.
}
\usage{
sio_df
}
\description{
Data included in the supermetroid package.
}
\seealso{
Other data: 
\code{\link{anatomecha_splits}},
\code{\link{src_run_df}}
}
\concept{data}
\concept{data
Splitsio data: time series observations

See vignette splits}
\keyword{datasets}
