% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sio_df}
\alias{sio_df}
\title{smdatasets}
\format{
An object of class \code{data.frame} with 19815 rows and 14 columns.
}
\usage{
sio_df
}
\description{
smdata included in the supermetroid package.
}
\seealso{
Other smdata: 
\code{\link{anatomecha_splits}},
\code{\link{distinct_routes}},
\code{\link{src_df}()}
}
\concept{smdata}
\concept{smdata
Splitsio smdata: time series observations

See vignette splits}
\keyword{datasets}
