% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sio_splits_df}
\alias{sio_splits_df}
\title{smdatasets}
\format{
An object of class \code{data.frame} with 19815 rows and 12 columns.
}
\usage{
sio_splits_df
}
\description{
smdata included in the supermetroid package.
}
\seealso{
Other smdata: 
\code{\link{anatomecha_splits}},
\code{\link{distinct_routes}},
\code{\link{from_to_df}},
\code{\link{sio_df}},
\code{\link{sio_routes_wide}},
\code{\link{src_df}()}
}
\concept{smdata}
\concept{smdata
Splits.io data for segments

Intermediary dataset from sio vignette.
Each row is a split time (segment) for a specific run, created in sio. Use
\link{sio_df} for analyses; this is the raw data before relabelling.}
\keyword{datasets}
