% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sio_splits_df}
\alias{sio_splits_df}
\title{Splits.io data for segments}
\format{
An object of class \code{data.frame} with 19815 rows and 12 columns.
}
\usage{
sio_splits_df
}
\description{
Intermediary dataset from sio vignette.
Each row is a split time (segment) for a specific run, created in sio. Use
sio_df for analyses; this is the raw data before relabelling.
}
\keyword{datasets}
